Update copyright, clean up README, add to .gitignore, remove .tool-versions (#192)
* docs(README.md): update title formatting and add website status badge
style(README.md): improve readability of local server URL
The title formatting was updated to use the more common Markdown syntax for headers, which improves readability and consistency with other Markdown files. A website status badge was added to provide quick access to the live site and indicate its current status. The local server URL was also updated to be a clickable link, improving usability for those reading the README.

* fix(default.html): update copyright year from 2023 to 2024
The copyright year was updated to reflect the current year, ensuring that the website's footer information is up-to-date.

* chore(.gitignore): ignore .tool-versions and .envrc files
The .tool-versions and .envrc files are added to the .gitignore file to prevent them from being tracked by Git. These files typically contain local development settings that should not be shared with the team or included in the repository.

* chore: remove .tool-versions file
The .tool-versions file was removed because it was no longer necessary. The project no longer relies on a specific version of Ruby, making the file redundant.