docs(coc.md): add Code of Conduct for Austin Hackers Association (#195)
Introduce a Code of Conduct document to establish guidelines for behavior
at Austin Hackers Association events. The document outlines expectations
for respectful interaction, privacy, accountability, and handling of
violations. It aims to create a fun, inclusive, and safe environment for
all participants by setting clear rules and consequences for misconduct.